move emulated run functions from tests to web
